+++ addenda98.go
@@ -24,7 +24,7 @@ type Addenda98 struct {
 	// OriginalTrace This field contains the Trace Number as originally included on the forward Entry or Prenotification.
 	// The RDFI must include the Original Entry Trace Number in the Addenda Record of an Entry being returned to an ODFI,
 	// in the Addenda Record of an 98, within an Acknowledgment Entry, or with an RDFI request for a copy of an authorization.
-	OriginalTrace int `json:"originalTrace"`
+	OriginalTrace string `json:"originalTrace"`
 	// OriginalDFI field contains the Receiving DFI Identification (addenda.RDFIIdentification) as originally included on the forward Entry or Prenotification that the RDFI is returning or correcting.
 	OriginalDFI string `json:"originalDFI"`
 	// CorrectedData
@@ -77,7 +77,7 @@ func (addenda98 *Addenda98) Parse(record string) {
 	// 4-6
 	addenda98.ChangeCode = record[3:6]
 	// 7-21
-	addenda98.OriginalTrace = addenda98.parseNumField(record[6:21])
+	addenda98.OriginalTrace = strings.TrimSpace(record[6:21])
 	// 28-35
 	addenda98.OriginalDFI = addenda98.parseStringField(record[27:35])
 	// 36-64
@@ -136,7 +136,7 @@ func (addenda98 *Addenda98) Validate() error {
 
 // OriginalTraceField returns a zero padded OriginalTrace string
 func (addenda98 *Addenda98) OriginalTraceField() string {
-	return addenda98.numericField(addenda98.OriginalTrace, 15)
+	return addenda98.stringField(addenda98.OriginalTrace, 15)
 }
 
 // OriginalDFIField returns a zero padded OriginalDFI string