+++ iatBatch_test.go
@@ -103,7 +103,7 @@ func mockInvalidAddenda17() *Addenda17 {
 func mockIATAddenda99() *Addenda99 {
 	addenda99 := NewAddenda99()
 	addenda99.ReturnCode = "R07"
-	addenda99.OriginalTrace = 231380100000001
+	addenda99.OriginalTrace = "231380100000001"
 	addenda99.OriginalDFI = "12104288"
 	addenda99.IATPaymentAmount("0000100000")
 	addenda99.IATAddendaInformation("Authorization Revoked")
@@ -113,7 +113,7 @@ func mockIATAddenda99() *Addenda99 {
 func mockIATAddenda98() *Addenda98 {
 	addenda98 := NewAddenda98()
 	addenda98.ChangeCode = "C01"
-	addenda98.OriginalTrace = 231380100000001
+	addenda98.OriginalTrace = "231380100000001"
 	addenda98.OriginalDFI = "12104288"
 	addenda98.CorrectedData = "89722-C3"
 	addenda98.TraceNumber = "121042880000001"